Study View Treatments API
This commit creates two new endpoints to support the new treatments
charts in the study view. It also adds additional logic to the
StudyViewFilterApplier to help it filter by treatment. The commit
also fixes two endpoints that 500ed when given a filter that results
in 0 samples.

`/treatments/patient` returns a list of PatientTreatmentRows. Each
patient treatment rows represents the number of patients that have
recieved a specific treatment.

`/treatments/sample` returns a list of SampleTreatmentRows. Each
sample treatment row represents either the number of samples
that were taken after a patient recieved a specific treatment,
or the number of samples that were taken before a patient recieved
a specific treatment.

The StudyViewFilter gained two new fields: sampleTreatmentFilters
and patientTreatmentFilters. These fileds correspond to the previously
described rows and filter out samples as you would expect.